+++ requirements.txt
@@ -3,3 +3,8 @@ opencv-python
 scikit-image
 # resolved as skimage
 pytesseract
+pygobject
+pycairo
+setuptools
+wheel
+weasyprint
