+++ testing/test_output_text/correct_text.txt
@@ -1,3 +1,4 @@
+# From Mac OS Preview Live Text OCR
 1
 New Career in
 Cyber... "Who Dis?"
