+++ testing/test_output_text/text_output_1_easyocr.txt
@@ -1,0 +1,44 @@
+1
+New Career in
+Cyber__ "Who Dis?"
+Cybersecurity is the hot new career field. It might even be hotter than the Kardashians
+Why; you ask? To put it simply;
+technology is spearheading global growth: With new
+technology such &s artificial
+and machine learning comes more cybersecurity
+demand: It's the whole supply and demand concept you learned in high school
+While more jobs are
+becoming remote; virtual, or replaced
+robots, how do you think
+are
+protected and secured? Cybersecurity is more than an IT helpdesk or the hackers
+you see on TV Itsa broad spectrum within all aspects of
+technology that has made
+cybersecurity
+an
+upward
+for careers, which will not
+anytime soon.
+What we know
+today will drastically change from what we know 1 year from now: We
+know that cybersecurity professionals make great money and get to wear hoodies and
+t0 work With all the positives do come some negatives You will learn about
+within this chapter:
+this
+you will have me; Jaclyn (Jax) Scott,
+you: You are encouraged
+to have a notebook close while
+this
+chapter and following chapters to help with
+Your data retention.
+intelligence =
+by -
+they :
+slope
+change
+Mip-E
+Mlops
+During 5
+chapter -
+leading `
+reading
