+++ testing/test_output_text/text_output_1_pytesseract.txt
