+++ testing/text_extractor_test.py
@@ -1,7 +1,5 @@
-import os
-import cv2
-import pytesseract
-import easyocr
+from keras_test import keras_ocr
+
 import numpy as np
 from PIL import Image
 
@@ -22,89 +20,12 @@ from PIL import Image
     Other OCRs:
         - EasyOCR
         - Keras-OCR
-        - DocTR
+        - Doc TR
 """
 
 
-def pytesseract_main():
-    # Determine the directory of the current script
-    script_dir = os.path.dirname(os.path.abspath(__file__))
-    # Relative path to the image file
-    image_file = os.path.join(script_dir, 'test_frame_1.jpg')
-
-    if os.path.exists(image_file):
-        # 1) Load the image
-        frame = cv2.imread(image_file)
-        
-        # 2) Try PIL Image
-        # image_pil = Image.open(image_file)
-        # frame = np.array(image_pil)
-        # if frame.dtype != np.uint8:
-        #     frame = frame.astype(np.uint8)
-        
-        # Check if the image was successfully loaded
-        if frame is not None:
-            # 1) Apply OCR to the image directly
-            text = pytesseract.image_to_string(frame)
-            
-            # 2) Try Preprocessing
-            # gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-            # gray = cv2.GaussianBlur(gray, (3, 3), 0)
-            # _, thresh = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY + cv2.THRESH_OTSU)
-            # text = pytesseract.image_to_string(thresh)
-            
-
-            # Save extracted text to a file
-            text_output_file = os.path.join(script_dir, 'text_output_1.txt')
-            with open(text_output_file, 'w') as file:
-                file.write(text)
-                
-            # Print extracted text
-            # print(f"Text extracted from {image_file}:")
-            # print(text)
-                
-            print(f"Extracted text saved to: {text_output_file}")
-        else:
-            print(f"CV2 failed to load image frame: {image_file}")
-    else:
-        print(f"File not found: {image_file}")
 
 
-def easyOCR_main():
-    # Determine the directory of the current script
-    script_dir = os.path.dirname(os.path.abspath(__file__))
-    # Relative path to the image file
-    image_file = os.path.join(script_dir, 'test_frame_1.jpg')
-
-    if os.path.exists(image_file):
-        # Load the image with OpenCV
-        frame = cv2.imread(image_file)
-        
-        # Check if the image was successfully loaded
-        if frame is not None:
-            # Initialize the EasyOCR Reader
-            reader = easyocr.Reader(['en'])  # Specify the language, 'en' for English
-            
-            # Convert the image to grayscale as EasyOCR may perform better on such images
-            gray_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-            
-            # Apply OCR directly to the grayscale image
-            results = reader.readtext(gray_frame)
-            
-            # Combine the extracted text into a single string
-            text = '\n'.join([result[1] for result in results])
-
-            # Save extracted text to a file
-            text_output_file = os.path.join(script_dir, 'text_output_1_easyocr.txt')
-            with open(text_output_file, 'w') as file:
-                file.write(text)
-                
-            print(f"Extracted text saved to: {text_output_file}")
-        else:
-            print(f"CV2 failed to load image frame: {image_file}")
-    else:
-        print(f"File not found: {image_file}")
-
-
+# MAIN
 if __name__ == "__main__":
-    easyOCR_main()
+    keras_ocr()
