--- testing/text_output_1_easyocr.txt
@@ -1,35 +1,0 @@
-New Career in
-Cyber__. "Who Dis?"
-Cybersecurity is the hot new career field. It might even be hotter than the Kardashians_
-Why; You ask? To put
-simply; technology is spearheading global growth: With new
-technology such as artificial intelligence and machine learning comgsomore cybersecurity
-demand: Its the whole supply and demand concept you learned in high school.
-While more jobs are becoming remote; virtual, or replaced by robots, how do you think
-they are protected and secured? Cybersecurity is more than an IT helpdesk or the hackers
-you see on TV: It$
-broad spectrum within all aspects of
-technology that has made
-cybersecurity an upward
-for careers, which will not
-anytime soon:
-What we know today will
-drastically change from what we know
-year from nOw: We
-know that cybersecurity professionals make great money and get to wear hoodies and
-flip
-to work: With all the positives do come some negatives You will learn about
-within this chapter:
-this
-chapter you will have me; Jaclyn (Jax) Scot;, leading you: You are encouraged
-to have
-notebook
-while reading this .
-and following chapters to help with
-Your data retention.
-slope
-change
--flops
-During
-close
-chapter !
